+++ src/main/java/calculator/service/CalculatorDelimiter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 import calculator.exception.InvalidInputException;
