+++ src/main/java/calculator/service/CalculatorManager.java
@@ -1,4 +1,6 @@
-package calculator;
+package calculator.service;
+
+import calculator.util.CalculatorParser;
 
 import java.util.List;
