+++ src/main/java/calculator/service/CalculatorNumberList.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 import calculator.exception.InvalidNumberException;
