+++ src/main/java/calculator/service/Delimiter.java
@@ -1,4 +1,6 @@
-package calculator;
+package calculator.service;
+
+import calculator.util.DefaultDelimiter;
 
 import java.util.List;
