+++ src/main/java/calculator/util/CalculatorParser.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.util;
 
 import calculator.exception.InvalidDelimiterException;
