+++ src/main/java/calculator/util/DefaultDelimiter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.util;
 
 import java.util.Arrays;
 import java.util.List;
