+++ src/main/java/calculator/view/CalculatorInput.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.view;
 
 import camp.nextstep.edu.missionutils.Console;
