+++ src/test/java/calculator/CalculatorDelimiterTest.java
@@ -1,10 +1,8 @@
 package calculator;
 
 import calculator.exception.InvalidInputException;
+import calculator.service.CalculatorDelimiter;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
