+++ src/test/java/calculator/CalculatorNumberListTest.java
@@ -1,6 +1,7 @@
 package calculator;
 
 import calculator.exception.InvalidNumberException;
+import calculator.service.CalculatorNumberList;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -9,7 +10,6 @@ import java.util.List;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class CalculatorNumberListTest {
